split functions.cpp into separate files
