+++ rmw_opensplice_cpp/CMakeLists.txt
@@ -25,7 +25,29 @@ ament_export_dependencies(rmw rosidl_generator_cpp rosidl_typesupport_opensplice
 ament_export_include_directories(${OpenSplice_INCLUDE_DIRS})
 
 link_directories(${OpenSplice_LIBRARY_DIRS})
-add_library(rmw_opensplice_cpp SHARED src/functions.cpp)
+add_library(rmw_opensplice_cpp SHARED
+  src/identifier.cpp
+  src/qos.cpp
+  src/rmw_client.cpp
+  src/rmw_compare_gids_equal.cpp
+  src/rmw_count.cpp
+  src/rmw_get_gid_for_publisher.cpp
+  src/rmw_get_implementation_identifier.cpp
+  src/rmw_guard_condition.cpp
+  src/rmw_init.cpp
+  src/rmw_node.cpp
+  src/rmw_publish.cpp
+  src/rmw_publisher.cpp
+  src/rmw_request.cpp
+  src/rmw_response.cpp
+  src/rmw_service.cpp
+  src/rmw_subscription.cpp
+  src/rmw_take.cpp
+  src/rmw_topic_names_and_types.cpp
+  src/rmw_trigger_guard_condition.cpp
+  src/rmw_wait.cpp
+  src/types.cpp
+)
 ament_target_dependencies(rmw_opensplice_cpp
   "rmw"
   "rosidl_generator_cpp"
