+++ rmw_opensplice_cpp/src/qos.hpp
@@ -1,0 +1,36 @@
+// Copyright 2014-2015 Open Source Robotics Foundation, Inc.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+#ifndef QOS_HPP_
+#define QOS_HPP_
+
+#include <ccpp_dds_dcps.h>
+#include <dds_dcps.h>
+#include <u_instanceHandle.h>
+
+#include "rmw/rmw.h"
+
+RMW_LOCAL
+bool get_datareader_qos(
+  DDS::Subscriber * subscriber,
+  const rmw_qos_profile_t & qos_profile,
+  DDS::DataReaderQos & datareader_qos);
+
+RMW_LOCAL
+bool get_datawriter_qos(
+  DDS::Publisher * publisher,
+  const rmw_qos_profile_t & qos_profile,
+  DDS::DataWriterQos & datawriter_qos);
+
+#endif  // QOS_HPP_
