+++ rmw_opensplice_cpp/src/rmw_client.cpp
@@ -1,0 +1,222 @@
+// Copyright 2014-2015 Open Source Robotics Foundation, Inc.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+#include <ccpp_dds_dcps.h>
+#include <dds_dcps.h>
+#include <string>
+
+#include "rosidl_typesupport_opensplice_cpp/identifier.hpp"
+
+#include "rmw/allocators.h"
+#include "rmw/error_handling.h"
+#include "rmw/impl/cpp/macros.hpp"
+#include "rmw/rmw.h"
+#include "rmw/types.h"
+
+
+#include "identifier.hpp"
+#include "qos.hpp"
+#include "types.hpp"
+
+// The extern "C" here enforces that overloading is not used.
+extern "C"
+{
+rmw_client_t *
+rmw_create_client(
+  const rmw_node_t * node,
+  const rosidl_service_type_support_t * type_support,
+  const char * service_name,
+  const rmw_qos_profile_t & qos_profile)
+{
+  if (!node) {
+    RMW_SET_ERROR_MSG("node handle is null");
+    return nullptr;
+  }
+  RMW_CHECK_TYPE_IDENTIFIERS_MATCH(
+    node handle,
+    node->implementation_identifier, opensplice_cpp_identifier,
+    return nullptr)
+
+  if (!type_support) {
+    RMW_SET_ERROR_MSG("type support handle is null");
+    return nullptr;
+  }
+  RMW_CHECK_TYPE_IDENTIFIERS_MATCH(
+    type support,
+    type_support->typesupport_identifier,
+    rosidl_typesupport_opensplice_cpp::typesupport_opensplice_identifier,
+    return nullptr)
+
+  auto node_info = static_cast<OpenSpliceStaticNodeInfo *>(node->data);
+  if (!node_info) {
+    RMW_SET_ERROR_MSG("node info handle is null");
+    return NULL;
+  }
+  auto participant = static_cast<DDS::DomainParticipant *>(node_info->participant);
+  if (!participant) {
+    RMW_SET_ERROR_MSG("participant handle is null");
+    return NULL;
+  }
+
+  const service_type_support_callbacks_t * callbacks =
+    static_cast<const service_type_support_callbacks_t *>(type_support->data);
+  if (!callbacks) {
+    RMW_SET_ERROR_MSG("callbacks handle is null");
+    return NULL;
+  }
+
+  DDS::DataReaderQos datareader_qos;
+  DDS::DataWriterQos datawriter_qos;
+
+  // Past this point, a failure results in unrolling code in the goto fail block.
+  rmw_client_t * client = nullptr;
+  const char * error_string = nullptr;
+  DDS::DataReader * response_datareader = nullptr;
+  DDS::ReadCondition * read_condition = nullptr;
+  void * requester = nullptr;
+  OpenSpliceStaticClientInfo * client_info = nullptr;
+  // Begin initializing elements.
+  client = rmw_client_allocate();
+  if (!client) {
+    RMW_SET_ERROR_MSG("failed to allocate client");
+    goto fail;
+  }
+
+  if (!get_datareader_qos(nullptr, qos_profile, datareader_qos)) {
+    goto fail;
+  }
+
+  if (!get_datawriter_qos(nullptr, qos_profile, datawriter_qos)) {
+    goto fail;
+  }
+
+  error_string = callbacks->create_requester(
+    participant, service_name,
+    reinterpret_cast<void **>(&requester),
+    reinterpret_cast<void **>(&response_datareader),
+    reinterpret_cast<const void *>(&datareader_qos),
+    reinterpret_cast<const void *>(&datawriter_qos),
+    &rmw_allocate);
+  if (error_string) {
+    RMW_SET_ERROR_MSG((std::string("failed to create requester: ") + error_string).c_str());
+    goto fail;
+  }
+  if (!requester) {
+    RMW_SET_ERROR_MSG("failed to create requester: requester is null");
+    goto fail;
+  }
+  if (!response_datareader) {
+    RMW_SET_ERROR_MSG("failed to create requester: response_datareader is null");
+    goto fail;
+  }
+
+  read_condition = response_datareader->create_readcondition(
+    DDS::ANY_SAMPLE_STATE, DDS::ANY_VIEW_STATE, DDS::ANY_INSTANCE_STATE);
+  if (!read_condition) {
+    RMW_SET_ERROR_MSG("failed to create read condition");
+    goto fail;
+  }
+
+  client_info = static_cast<OpenSpliceStaticClientInfo *>(
+    rmw_allocate(sizeof(OpenSpliceStaticClientInfo)));
+  if (!client_info) {
+    RMW_SET_ERROR_MSG("failed to allocate memory");
+    goto fail;
+  }
+  client_info->requester_ = requester;
+  client_info->callbacks_ = callbacks;
+  client_info->response_datareader_ = response_datareader;
+  client_info->read_condition_ = read_condition;
+
+  client->implementation_identifier = opensplice_cpp_identifier;
+  client->data = client_info;
+  return client;
+fail:
+  if (response_datareader) {
+    if (read_condition) {
+      if (response_datareader->delete_readcondition(read_condition) != DDS::RETCODE_OK) {
+        fprintf(stderr, "leaking readcondition while handling failure\n");
+      }
+    }
+  }
+
+  if (requester) {
+    const char * error_string = callbacks->destroy_requester(requester, &rmw_free);
+    if (error_string) {
+      std::stringstream ss;
+      ss << "failed to destroy requester: " << error_string << ", at: " <<
+        __FILE__ << ":" << __LINE__ << '\n';
+      (std::cerr << ss.str()).flush();
+    }
+  }
+  if (client_info) {
+    rmw_free(client_info);
+  }
+  if (client) {
+    rmw_client_free(client);
+  }
+  return nullptr;
+}
+
+rmw_ret_t
+rmw_destroy_client(rmw_client_t * client)
+{
+  if (!client) {
+    RMW_SET_ERROR_MSG("client handle is null");
+    return RMW_RET_ERROR;
+  }
+  RMW_CHECK_TYPE_IDENTIFIERS_MATCH(
+    client handle,
+    client->implementation_identifier, opensplice_cpp_identifier,
+    return RMW_RET_ERROR)
+
+  OpenSpliceStaticClientInfo * client_info =
+    static_cast<OpenSpliceStaticClientInfo *>(client->data);
+
+  auto result = RMW_RET_OK;
+  if (client_info) {
+    auto response_datareader = client_info->response_datareader_;
+    if (response_datareader) {
+      auto read_condition = client_info->read_condition_;
+      if (read_condition) {
+        if (response_datareader->delete_readcondition(read_condition) != DDS::RETCODE_OK) {
+          RMW_SET_ERROR_MSG("failed to delete readcondition");
+          result = RMW_RET_ERROR;
+        }
+        client_info->read_condition_ = nullptr;
+      }
+    }
+  } else {
+    RMW_SET_ERROR_MSG("client_info handle is null");
+    return RMW_RET_ERROR;
+  }
+
+  const service_type_support_callbacks_t * callbacks =
+    static_cast<const service_type_support_callbacks_t *>(client_info->callbacks_);
+  if (!callbacks) {
+    RMW_SET_ERROR_MSG("callbacks handle is null");
+    return RMW_RET_ERROR;
+  }
+
+  const char * error_string = callbacks->destroy_requester(client_info->requester_, &rmw_free);
+  if (error_string) {
+    RMW_SET_ERROR_MSG((std::string("failed to destroy requester: ") + error_string).c_str());
+    return RMW_RET_ERROR;
+  }
+
+  rmw_free(client_info);
+  rmw_client_free(client);
+  return result;
+}
+}  // extern "C"
