+++ rmw_opensplice_cpp/src/rmw_compare_gids_equal.cpp
@@ -1,0 +1,63 @@
+// Copyright 2014-2015 Open Source Robotics Foundation, Inc.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+#include "rmw/allocators.h"
+#include "rmw/error_handling.h"
+#include "rmw/impl/cpp/macros.hpp"
+#include "rmw/rmw.h"
+#include "rmw/types.h"
+
+#include "identifier.hpp"
+#include "types.hpp"
+
+// The extern "C" here enforces that overloading is not used.
+extern "C"
+{
+rmw_ret_t
+rmw_compare_gids_equal(const rmw_gid_t * gid1, const rmw_gid_t * gid2, bool * result)
+{
+  if (!gid1) {
+    RMW_SET_ERROR_MSG("gid1 is null");
+    return RMW_RET_ERROR;
+  }
+  RMW_CHECK_TYPE_IDENTIFIERS_MATCH(
+    gid1,
+    gid1->implementation_identifier, opensplice_cpp_identifier,
+    return RMW_RET_ERROR)
+  if (!gid2) {
+    RMW_SET_ERROR_MSG("gid2 is null");
+    return RMW_RET_ERROR;
+  }
+  RMW_CHECK_TYPE_IDENTIFIERS_MATCH(
+    gid2,
+    gid2->implementation_identifier, opensplice_cpp_identifier,
+    return RMW_RET_ERROR)
+  if (!result) {
+    RMW_SET_ERROR_MSG("result is null");
+    return RMW_RET_ERROR;
+  }
+  auto detail1 = reinterpret_cast<const OpenSplicePublisherGID *>(gid1->data);
+  if (!detail1) {
+    RMW_SET_ERROR_MSG("gid1 is invalid");
+    return RMW_RET_ERROR;
+  }
+  auto detail2 = reinterpret_cast<const OpenSplicePublisherGID *>(gid2->data);
+  if (!detail2) {
+    RMW_SET_ERROR_MSG("gid2 is invalid");
+    return RMW_RET_ERROR;
+  }
+  *result = detail1->publication_handle == detail2->publication_handle;
+  return RMW_RET_OK;
+}
+}  // extern "C"
