+++ rmw_opensplice_cpp/src/rmw_get_implementation_identifier.cpp
@@ -1,0 +1,27 @@
+// Copyright 2014-2015 Open Source Robotics Foundation, Inc.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+#include "rmw/rmw.h"
+
+#include "identifier.hpp"
+
+// The extern "C" here enforces that overloading is not used.
+extern "C"
+{
+const char *
+rmw_get_implementation_identifier()
+{
+  return opensplice_cpp_identifier;
+}
+}  // extern "C"
