+++ rmw_opensplice_cpp/src/rmw_init.cpp
@@ -1,0 +1,35 @@
+// Copyright 2014-2015 Open Source Robotics Foundation, Inc.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+#include <ccpp_dds_dcps.h>
+#include <dds_dcps.h>
+
+#include "rmw/error_handling.h"
+#include "rmw/rmw.h"
+#include "rmw/types.h"
+
+// The extern "C" here enforces that overloading is not used.
+extern "C"
+{
+rmw_ret_t
+rmw_init()
+{
+  DDS::DomainParticipantFactory_var dp_factory = DDS::DomainParticipantFactory::get_instance();
+  if (!dp_factory) {
+    RMW_SET_ERROR_MSG("failed to get domain participant factory");
+    return RMW_RET_ERROR;
+  }
+  return RMW_RET_OK;
+}
+}  // extern "C"
