+++ rmw_opensplice_cpp/src/rmw_publish.cpp
@@ -1,0 +1,65 @@
+// Copyright 2014-2015 Open Source Robotics Foundation, Inc.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+#include <string>
+
+#include "rmw/allocators.h"
+#include "rmw/error_handling.h"
+#include "rmw/impl/cpp/macros.hpp"
+#include "rmw/rmw.h"
+#include "rmw/types.h"
+
+#include "identifier.hpp"
+#include "types.hpp"
+
+// The extern "C" here enforces that overloading is not used.
+extern "C"
+{
+rmw_ret_t
+rmw_publish(const rmw_publisher_t * publisher, const void * ros_message)
+{
+  if (!publisher) {
+    RMW_SET_ERROR_MSG("publisher handle is null");
+    return RMW_RET_ERROR;
+  }
+  RMW_CHECK_TYPE_IDENTIFIERS_MATCH(
+    publisher handle,
+    publisher->implementation_identifier, opensplice_cpp_identifier,
+    return RMW_RET_ERROR)
+  if (!ros_message) {
+    RMW_SET_ERROR_MSG("ros message handle is null");
+    return RMW_RET_ERROR;
+  }
+
+  const OpenSpliceStaticPublisherInfo * publisher_info =
+    static_cast<const OpenSpliceStaticPublisherInfo *>(publisher->data);
+  if (!publisher_info) {
+    RMW_SET_ERROR_MSG("publisher info handle is null");
+    return RMW_RET_ERROR;
+  }
+  DDS::DataWriter * topic_writer = publisher_info->topic_writer;
+  const message_type_support_callbacks_t * callbacks = publisher_info->callbacks;
+  if (!callbacks) {
+    RMW_SET_ERROR_MSG("callbacks handle is null");
+    return RMW_RET_ERROR;
+  }
+
+  const char * error_string = callbacks->publish(topic_writer, ros_message);
+  if (error_string) {
+    RMW_SET_ERROR_MSG((std::string("failed to publish:") + error_string).c_str());
+    return RMW_RET_ERROR;
+  }
+  return RMW_RET_OK;
+}
+}  // extern "C"
