+++ rmw_opensplice_cpp/src/rmw_take.cpp
@@ -1,0 +1,115 @@
+// Copyright 2014-2015 Open Source Robotics Foundation, Inc.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+#include "rmw/allocators.h"
+#include "rmw/error_handling.h"
+#include "rmw/impl/cpp/macros.hpp"
+#include "rmw/rmw.h"
+#include "rmw/types.h"
+
+#include "identifier.hpp"
+#include "types.hpp"
+
+RMW_LOCAL
+rmw_ret_t
+take(
+  const rmw_subscription_t * subscription, void * ros_message, bool * taken,
+  DDS::InstanceHandle_t * sending_publication_handle)
+{
+  // Note: gid is allowed to be nullptr, if unused.
+  if (!subscription) {
+    RMW_SET_ERROR_MSG("subscription handle is null");
+    return RMW_RET_ERROR;
+  }
+  RMW_CHECK_TYPE_IDENTIFIERS_MATCH(
+    subscription handle,
+    subscription->implementation_identifier, opensplice_cpp_identifier,
+    return RMW_RET_ERROR)
+
+  if (ros_message == nullptr) {
+    RMW_SET_ERROR_MSG("ros_message argument cannot be null");
+    return RMW_RET_ERROR;
+  }
+
+  if (taken == nullptr) {
+    RMW_SET_ERROR_MSG("taken argument cannot be null");
+    return RMW_RET_ERROR;
+  }
+
+  OpenSpliceStaticSubscriberInfo * subscriber_info =
+    static_cast<OpenSpliceStaticSubscriberInfo *>(subscription->data);
+  if (!subscriber_info) {
+    RMW_SET_ERROR_MSG("subscriber info handle is null");
+    return RMW_RET_ERROR;
+  }
+  DDS::DataReader * topic_reader = subscriber_info->topic_reader;
+  if (!topic_reader) {
+    RMW_SET_ERROR_MSG("topic reader handle is null");
+    return RMW_RET_ERROR;
+  }
+  const message_type_support_callbacks_t * callbacks = subscriber_info->callbacks;
+  if (!callbacks) {
+    RMW_SET_ERROR_MSG("callbacks handle is null");
+    return RMW_RET_ERROR;
+  }
+
+  const char * error_string = callbacks->take(
+    topic_reader,
+    subscriber_info->ignore_local_publications,
+    ros_message, taken, sending_publication_handle);
+  // If no data was taken, that's not captured as an error here, but instead taken is set to false.
+  if (error_string) {
+    RMW_SET_ERROR_MSG((std::string("failed to take: ") + error_string).c_str());
+    return RMW_RET_ERROR;
+  }
+
+  return RMW_RET_OK;
+}
+
+// The extern "C" here enforces that overloading is not used.
+extern "C"
+{
+rmw_ret_t
+rmw_take(const rmw_subscription_t * subscription, void * ros_message, bool * taken)
+{
+  return take(subscription, ros_message, taken, nullptr);
+}
+
+rmw_ret_t
+rmw_take_with_info(
+  const rmw_subscription_t * subscription,
+  void * ros_message,
+  bool * taken,
+  rmw_message_info_t * message_info)
+{
+  if (!message_info) {
+    RMW_SET_ERROR_MSG("message info is null");
+    return RMW_RET_ERROR;
+  }
+  DDS::InstanceHandle_t sending_publication_handle;
+  auto ret = take(subscription, ros_message, taken, &sending_publication_handle);
+  if (ret != RMW_RET_OK) {
+    // Error string is already set.
+    return RMW_RET_ERROR;
+  }
+
+  rmw_gid_t * sender_gid = &message_info->publisher_gid;
+  sender_gid->implementation_identifier = opensplice_cpp_identifier;
+  memset(sender_gid->data, 0, RMW_GID_STORAGE_SIZE);
+  auto detail = reinterpret_cast<OpenSplicePublisherGID *>(sender_gid->data);
+  detail->publication_handle = sending_publication_handle;
+
+  return RMW_RET_OK;
+}
+}  // extern "C"
