+++ rmw_opensplice_cpp/src/rmw_topic_names_and_types.cpp
@@ -1,0 +1,182 @@
+// Copyright 2014-2015 Open Source Robotics Foundation, Inc.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+#include <map>
+#include <set>
+#include <string>
+
+#include "rmw/allocators.h"
+#include "rmw/error_handling.h"
+#include "rmw/rmw.h"
+#include "rmw/types.h"
+
+#include "identifier.hpp"
+#include "types.hpp"
+
+RMW_LOCAL
+void
+destroy_topic_names_and_types(
+  rmw_topic_names_and_types_t * topic_names_and_types)
+{
+  if (topic_names_and_types->topic_count) {
+    for (size_t i = 0; i < topic_names_and_types->topic_count; ++i) {
+      delete topic_names_and_types->topic_names[i];
+      delete topic_names_and_types->type_names[i];
+      topic_names_and_types->topic_names[i] = nullptr;
+      topic_names_and_types->type_names[i] = nullptr;
+    }
+    if (topic_names_and_types->topic_names) {
+      rmw_free(topic_names_and_types->topic_names);
+      topic_names_and_types->topic_names = nullptr;
+    }
+    if (topic_names_and_types->type_names) {
+      rmw_free(topic_names_and_types->type_names);
+      topic_names_and_types->type_names = nullptr;
+    }
+    topic_names_and_types->topic_count = 0;
+  }
+}
+
+// The extern "C" here enforces that overloading is not used.
+extern "C"
+{
+rmw_ret_t
+rmw_get_topic_names_and_types(
+  const rmw_node_t * node,
+  rmw_topic_names_and_types_t * topic_names_and_types)
+{
+  if (!node) {
+    RMW_SET_ERROR_MSG("node handle is null");
+    return RMW_RET_ERROR;
+  }
+  if (node->implementation_identifier != opensplice_cpp_identifier) {
+    RMW_SET_ERROR_MSG("node handle is not from this rmw implementation");
+    return RMW_RET_ERROR;
+  }
+  if (!topic_names_and_types) {
+    RMW_SET_ERROR_MSG("topics handle is null");
+    return RMW_RET_ERROR;
+  }
+  if (topic_names_and_types->topic_count) {
+    RMW_SET_ERROR_MSG("topic count is not zero");
+    return RMW_RET_ERROR;
+  }
+  if (topic_names_and_types->topic_names) {
+    RMW_SET_ERROR_MSG("topic names is not null");
+    return RMW_RET_ERROR;
+  }
+  if (topic_names_and_types->type_names) {
+    RMW_SET_ERROR_MSG("type names is not null");
+    return RMW_RET_ERROR;
+  }
+
+  auto node_info = static_cast<OpenSpliceStaticNodeInfo *>(node->data);
+  if (!node_info) {
+    RMW_SET_ERROR_MSG("node info handle is null");
+    return RMW_RET_ERROR;
+  }
+  if (!node_info->publisher_listener) {
+    RMW_SET_ERROR_MSG("publisher listener handle is null");
+    return RMW_RET_ERROR;
+  }
+  if (!node_info->subscriber_listener) {
+    RMW_SET_ERROR_MSG("subscriber listener handle is null");
+    return RMW_RET_ERROR;
+  }
+
+  // combine publisher and subscriber information
+  std::map<std::string, std::set<std::string>> topics_with_multiple_types;
+  for (auto it : node_info->publisher_listener->topic_names_and_types) {
+    for (auto & jt : it.second) {
+      topics_with_multiple_types[it.first].insert(jt);
+    }
+  }
+  for (auto it : node_info->subscriber_listener->topic_names_and_types) {
+    for (auto & jt : it.second) {
+      topics_with_multiple_types[it.first].insert(jt);
+    }
+  }
+
+  // ignore inconsistent types
+  std::map<std::string, std::string> topics;
+  for (auto & it : topics_with_multiple_types) {
+    if (it.second.size() != 1) {
+      fprintf(stderr, "topic type mismatch - ignoring topic '%s'\n", it.first.c_str());
+      continue;
+    }
+    topics[it.first] = *it.second.begin();
+  }
+
+  // reformat type name
+  std::string substr = "::msg::dds_::";
+  for (auto & it : topics) {
+    size_t substr_pos = it.second.find(substr);
+    if (it.second[it.second.size() - 1] == '_' && substr_pos != std::string::npos) {
+      it.second = it.second.substr(0, substr_pos) + "/" + it.second.substr(
+        substr_pos + substr.size(), it.second.size() - substr_pos - substr.size() - 1);
+    }
+  }
+
+  // copy data into result handle
+  if (topics.size() > 0) {
+    topic_names_and_types->topic_names = static_cast<char **>(
+      rmw_allocate(sizeof(char *) * topics.size()));
+    if (!topic_names_and_types->topic_names) {
+      RMW_SET_ERROR_MSG("failed to allocate memory for topic names")
+      return RMW_RET_ERROR;
+    }
+    topic_names_and_types->type_names = static_cast<char **>(
+      rmw_allocate(sizeof(char *) * topics.size()));
+    if (!topic_names_and_types->type_names) {
+      rmw_free(topic_names_and_types->topic_names);
+      RMW_SET_ERROR_MSG("failed to allocate memory for type names")
+      return RMW_RET_ERROR;
+    }
+    for (auto it : topics) {
+      char * topic_name = strdup(it.first.c_str());
+      if (!topic_name) {
+        RMW_SET_ERROR_MSG("failed to allocate memory for topic name")
+        goto fail;
+      }
+      char * type_name = strdup(it.second.c_str());
+      if (!type_name) {
+        rmw_free(topic_name);
+        RMW_SET_ERROR_MSG("failed to allocate memory for type name")
+        goto fail;
+      }
+      size_t i = topic_names_and_types->topic_count;
+      topic_names_and_types->topic_names[i] = topic_name;
+      topic_names_and_types->type_names[i] = type_name;
+      ++topic_names_and_types->topic_count;
+    }
+  }
+
+  return RMW_RET_OK;
+fail:
+  destroy_topic_names_and_types(topic_names_and_types);
+  return RMW_RET_ERROR;
+}
+
+rmw_ret_t
+rmw_destroy_topic_names_and_types(
+  rmw_topic_names_and_types_t * topic_names_and_types)
+{
+  if (!topic_names_and_types) {
+    RMW_SET_ERROR_MSG("topics handle is null");
+    return RMW_RET_ERROR;
+  }
+  destroy_topic_names_and_types(topic_names_and_types);
+  return RMW_RET_OK;
+}
+}  // extern "C"
